Add safety checks for adding/deleting devices with no hostname

+++ app/Observers/DeviceObserver.php
@@ -10,6 +10,7 @@ use App\Models\Eventlog;
 use File;
 use Illuminate\Support\Facades\App;
 use LibreNMS\Enum\Severity;
+use LibreNMS\Exceptions\HostNameEmptyException;
 use LibreNMS\Exceptions\HostRenameException;
 use Log;
 
@@ -92,19 +93,21 @@ class DeviceObserver
      */
     public function deleted(Device $device): void
     {
-        // delete rrd files
-        $host_dir = Rrd::dirFromHost($device->hostname);
-        try {
-            $result = File::deleteDirectory($host_dir);
+        if (!empty($device->hostname)) {
+            // delete rrd files
+            $host_dir = Rrd::dirFromHost($device->hostname);
+            try {
+                $result = File::deleteDirectory($host_dir);
 
-            if (! $result) {
-                Log::debug("Could not delete RRD files for: $device->hostname");
+                if (! $result) {
+                    Log::debug("Could not delete RRD files for: $device->hostname");
+                }
+            } catch (\Exception $e) {
+                Log::error("Could not delete RRD files for: $device->hostname", [$e]);
             }
-        } catch (\Exception $e) {
-            Log::error("Could not delete RRD files for: $device->hostname", [$e]);
         }
 
-        Eventlog::log("Device $device->hostname has been removed", 0, 'system', Severity::Notice);
+        Eventlog::log("Device " . ($device->hostname ?: $device->device_id) . " has been removed", 0, 'system', Severity::Notice);
 
         (new Oxidized)->reloadNodes();
     }
